Revert "bumped libbuildpack-dynatrace to v1.5.0"
