--- fixtures/fake_dynatrace_api/fake_config.json
@@ -1,11 +1,0 @@
-{
-  "revision":1234567890,
-  "properties":
-  [
-    {
-      "section":"fakesection",
-      "key":"fakekey",
-      "value":"fakevalue"
-    }
-  ]
-}
