--- fixtures/fake_dynatrace_api/ruxitagentproc.conf
@@ -1,3 +1,0 @@
-# some comment
-[testsection]
-testkey testvalue
