+++ fixtures/util/dynatrace/README.md
@@ -1,0 +1,4 @@
+# TODO:
+Get context on how this app was created. See
+https://github.com/cloudfoundry/dotnet-core-buildpack/commits/master/fixtures/fake_dynatrace_api
+for original app history.
