+++ fixtures/util/dynatrace/dynatrace-env.sh
