+++ fixtures/util/dynatrace/liboneagentproc.so
