+++ fixtures/util/dynatrace/manifest.json
